SchedulePlanner.App

Description

Sort of the Main script for all the common operations, independant of the program instance (webservice, command line)

Documentation

reportAndPrint :: Text -> Bool -> Text -> IO ()

perform the calculation and print the result to the command line

reportAndExecute :: Text -> Bool -> DataFile -> Writer Text ()

Evaluates the transformed json, compiles (useful) error messages, runs the algorithm and returns a writer of any output created.