SchedulePlanner.Serialize

Description

Hold the capeablilities to get and export in- and output data as well as (de)serialize it

Documentation

ruleKey :: Text

Key for the rules data in the json input

lessonKey :: Text

Key for the lesson data in the json input

scopeKey :: Text

Key for the scope property in Rule objects in the json input

severityKey :: Text

Key for the severity property in Rule objects in the json input

ruleDayKey :: Text

Key for the day property in Rule objects in the json input

ruleSlotKey :: Text

Key for the slot property in Rule objects in the json input

subjectKey :: Text

Key for the subject property in Lesson objects in the json input

lessonDayKey :: Text

Key for the day property in Lesson objects in the json input

lessonSlotKey :: Text

Key for the slot property in Rule objects in the json input

daysPerWeek :: Int

How many days a week has

slotsPerDay :: Int

The amount of imeslots each day

cellWidth :: Int

The caracter width of a single slot in output

data DataFile

Constructors

DataFile [Rule] [Lesson Text]

mapToJSON :: ToJSON a => Map Text a -> Value

Convert a suitable Map to a JSON Value

getFromFile :: FilePath -> IO (Maybe DataFile)

Open a file and return the contents as parsed json

writeToFile :: ToJSON a => FilePath -> Map Text a -> IO ()

Open a file and write json to it

shortSubject :: Show s => s -> String

formatSchedule :: Show s => MappedSchedule s -> Text

Transform a MappedSchedule into a printable, and more importantly, readable Text