Main

Description

main function and outside communication for this software. This module takes care of reading all input and checking for correctness as well as providing useful feedback upon encountering errors.

Documentation

stdFileName :: [Char]

Legacy hard coded name of inputfile

ruleKey :: [Char]

Key for the rules data in the json input

lessonKey :: [Char]

Key for the lesson data in the json input

scopeKey :: [Char]

Key for the scope property in Rule objects in the json input

severityKey :: [Char]

Key for the severity property in Rule objects in the json input

ruleDayKey :: [Char]

Key for the day property in Rule objects in the json input

ruleSlotKey :: [Char]

Key for the slot property in Rule objects in the json input

subjectKey :: [Char]

Key for the subject property in Lesson objects in the json input

lessonDayKey :: [Char]

Key for the day property in Lesson objects in the json input

lessonSlotKey :: [Char]

Key for the slot property in Rule objects in the json input

testLessons :: [Lesson]

Legacy test data

putErrorLine :: String -> IO ()

Print a line to stdout

getFromFile :: JSON a => String -> IO (Result a)

Open a file and return the contents as parsed json

toNative :: Result JSValue -> Result ([Result Rule], [Result Lesson])

Turns parsed json values into the internally used datastructures.

extractRules :: JSValue -> Result [Result Rule]

Turns a parsed json value into a List of Rules or return an Error

extractLessons :: JSValue -> Result [Result Lesson]

Turns a parsed json value into a List of Lessons or return an Error

reportAndExecute :: Result ([Result Rule], [Result Lesson]) -> IO ()

Evaluates the transformed json, compiles (useful) error messages, prints them and then runs the algorithm or, if the errors are too severe, abourts.

main :: IO ()

main function. Handles reading command line arguments, the json input and starting execution.