Main

Description

main function and outside communication for this software. This module takes care of reading all input and checking for correctness as well as providing useful feedback upon encountering errors.

Documentation

main :: IO ()

main function. Handles reading command line arguments, the json input and starts execution.