SchedulePlanner.App

Documentation

reportAndPrint :: Text -> Bool -> Text -> IO ()

reportAndExecute :: Text -> Bool -> DataFile -> Writer Text ()

Evaluates the transformed json, compiles (useful) error messages, prints them and then runs the algorithm or, if the errors are too severe, aborts.